% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-getClinDT.R
\name{checkVarInData}
\alias{checkVarInData}
\title{Check if specified variables are in the data.
If they are not, they are removed from specified
variables and a message is printed.}
\usage{
checkVarInData(var, data, label)
}
\arguments{
\item{var}{Character vector with variables.}

\item{data}{Data.frame with data.}

\item{label}{String with label used in message.}
}
\value{
\code{var} present in \code{data},
NULL if empty
}
\description{
Check if specified variables are in the data.
If they are not, they are removed from specified
variables and a message is printed.
}
