% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-palettes.R
\name{clinColors}
\alias{clinColors}
\title{Colors of 'clinUtils'}
\usage{
clinColors(n, alpha = 1, begin = 0, end = 1, direction = 1, option = "D")
}
\arguments{
\item{n}{The number of colors (\eqn{\ge 1}) to be in the palette.}

\item{alpha}{The alpha transparency, a number in [0,1], see argument alpha in
\code{\link[grDevices]{hsv}}.}

\item{begin}{The (corrected) hue in [0,1] at which the viridis colormap begins.}

\item{end}{The (corrected) hue in [0,1] at which the viridis colormap ends.}

\item{direction}{Sets the order of colors in the scale. If 1, the default, colors
are ordered from darkest to lightest. If -1, the order of colors is reversed.}

\item{option}{A character string indicating the colormap option to use. Four
options are available: "magma" (or "A"), "inferno" (or "B"), "plasma" (or "C"),
"viridis" (or "D", the default option) and "cividis" (or "E").}
}
\description{
Default color palette is the color-blind viridis palette.
See documentation of \code{\link[viridisLite]{viridis}}.
}
