% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-palettes.R
\docType{data}
\name{clinLinetypes}
\alias{clinLinetypes}
\title{Linetypes of 'clinUtils'}
\format{
An object of class \code{character} of length 6.
}
\usage{
clinLinetypes
}
\value{
A character vector of linetypes.
}
\description{
A set of default linetypes are available as a vector.
}
\keyword{datasets}
