% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-palettes.R
\docType{data}
\name{clinShapes}
\alias{clinShapes}
\title{Shapes of 'clinUtils'}
\format{
An object of class \code{integer} of length 24.
}
\usage{
clinShapes
}
\description{
A set of default shapes are available as named vector.
}
\keyword{datasets}
