% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-palettes.R
\docType{data}
\name{clinShapesText}
\alias{clinShapesText}
\title{Shapes of 'clinUtils' as text}
\format{
An object of class \code{character} of length 24.
}
\usage{
clinShapesText
}
\description{
A set of default shapes are available as character vector.
}
\keyword{datasets}
