% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-palettes.R
\name{clinUtils-palette}
\alias{clinUtils-palette}
\title{Parameters for all palette functions for clinical visualizations.}
\arguments{
\item{n}{Integer of length 1, number of elements in palette.}

\item{x}{Vector with elements used for palette.
If factor, the levels are used, otherwise the unique elements of the vector.
Missing values are automatically removed, excepted if
\code{includeNA} is set to TRUE.}

\item{includeNA}{Logical (FALSE by default), 
should NA elements be retained in the palette in case
\code{x} is specified?}
}
\value{
Specific palettes used in \code{clinUtils}.
}
\description{
Parameters for all palette functions for clinical visualizations.
}
