% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-palettesCDISC.R
\docType{data}
\name{colorPaletteNRIND}
\alias{colorPaletteNRIND}
\title{Color palette for a standard CDISC Normal/Reference
Range Indicator.}
\format{
A named character vector with color for 
typical Normal Reference Range Indicator variable:
\itemize{
\item "LOW": orange
\item "NORMAL": green4
\item "HIGH": orange
\item "ABNORMAL": red
\item "UNKNOWN" or 'NA': grey
\item "NA": grey
}
}
\usage{
colorPaletteNRIND
}
\description{
Color palette for a standard CDISC Normal/Reference
Range Indicator.
}
\keyword{datasets}
