% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareFigs.R
\name{compareFigOne}
\alias{compareFigOne}
\title{Compare two figures to each other}
\usage{
compareFigOne(pathFigRef, pathFigNew, outputDir = "./compareFigs")
}
\arguments{
\item{pathFigRef}{String with path to the reference figure (only one)}

\item{pathFigNew}{String with path to the new figure (only one)}

\item{outputDir}{String with path to the folder where figure with difference are saved (if any)
as produced with 'ImageMagick compare' command.}
}
\value{
No returned value, message is printed in the current
console with 'ok!' if no difference, otherwise: 'DIFFERENT!'
}
\description{
If the figures are in PNG format, raster image are checked if they are identifical.
If ImageMagick is available (and only in case the figures differ for PNG), the figure with
the difference is saved in the \code{outputDir} folder, under the name '[filename]-diff.png'
}
\author{
Laure Cougnaud
}
