% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareFigs.R
\name{compareFigs}
\alias{compareFigs}
\title{Compare reference with new figures in png format with similar names.}
\usage{
compareFigs(pathFigRef, pathFigNew, outputDir = "./compareFigs", ...)
}
\arguments{
\item{pathFigRef}{String with path of directory where the reference figures have been stored,
or directly the paths of the reference files.}

\item{pathFigNew}{String with path of directory where the new figures have been stored,
or directly the paths of the new files.}

\item{outputDir}{String with path to the folder where figure with difference are saved (if any)
as produced with 'ImageMagick compare' command.}

\item{...}{Additional parameters for \code{\link{list.files}} function,
e.g. pattern of files to consider}
}
\value{
No returned value, message is printed in the current
console with 'ok!' if no difference, otherwise: 'DIFFERENT!'
}
\description{
The figures are mapped based on filename and only the 
figures present both in \code{pathFigRef} and \code{pathFigNew}
are considered.
}
\details{
If \href{https://imagemagick.org/}{ImageMagick} is available, in case the figures differ, the figures with
the difference is saved in the \code{outputDir} folder, under the name '[filename]-diff.png'
}
\examples{
\dontrun{

# ImageMagick should be installed and available
data(dataADaMCDISCP01)

library(ggplot2)

dataLB <- dataADaMCDISCP01$ADLBC
tmpdir <- tempdir()

# default ggplot2 behaviour
pathFigRef <- file.path(tmpdir, "compareFigs_ref")
dir.create(pathFigRef)
gg1 <- ggplot(data = subset(dataLB, PARAMCD == "CHOL"), 
	aes(x = ADY, y = CHG, color = SUBJID)) +
	geom_line(aes(linetype = SUBJID)) +
	geom_point(aes(shape = SUBJID)) +
	theme(legend.position = "bottom") +
	ggtitle("Patient profiles")
ggsave(filename = file.path(pathFigRef, "example1.png"), plot = gg1)
ggsave(filename = file.path(pathFigRef, "example2.png"), plot = gg1)
ggsave(filename = file.path(pathFigRef, "example3.png"), plot = gg1)

# use default-color, and linetype/shape palette with enough elements:
pathFigNew <- file.path(tmpdir, "compareFigs_new")
dir.create(pathFigNew)
ggsave(
	filename = file.path(pathFigNew, "example1.png"), 
	plot = gg1
)
ggsave(
	filename = file.path(pathFigNew, "example2.png"), 
	plot = gg1 + ggtitle("This is a new title")
)
ggsave(
	filename = file.path(pathFigNew, "example3.png"), 
	plot = gg1 + scale_color_manual(values = getColorPalette(x = dataLB$SUBJID))
)

compareFigs(pathFigRef = pathFigRef, pathFigNew = pathFigNew)

}
}
\author{
Laure Cougnaud
}
