% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareTables.R
\name{convertToDatatable}
\alias{convertToDatatable}
\title{Convert to data.table}
\usage{
convertToDatatable(data)
}
\arguments{
\item{data}{A data.frame}
}
\value{
A \code{data.table} object.
}
\description{
Convert a data frame into a \code{data.table} object.
}
