% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataSDTMCDISCP01}
\alias{dataSDTMCDISCP01}
\title{Example of SDTM datasets from the CDISC original Pilot 01 study}
\format{
List of data.frames containing the SDTM
dataset for each selected domain.\cr
Labels for the different variables across datasets
is available via the \code{labelVars} attribute.
}
\source{
Original (and entire) datasets are available in: \cr
\url{https://github.com/phuse-org/phuse-scripts/tree/master/data/sdtm/cdiscpilot01}
See in particular the \emph{define.xml} file
for further description of the
datasets and variables name.
}
\description{
This contains a subset of the CDISC original Pilot 01 study dataset for:
\itemize{
\item{a selected subset of subjects}
\item{a selected subset of domains: 
\itemize{
\item{adverse event ('ae')}
\item{concomitant medications ('cm')}
\item{disposition ('ds')}
\item{demographics ('dm')}
\item{treatment exposure ('ex')}
\item{laboratory ('lb')}
\item{medical history ('mh')}
\item{questionnaire ('qs') - only:
\itemize{
\item ADAS-Cog (11) primary endpoint (\code{QSTESTCD == 'ACTOT'})
\item CIBIC+ primary endpoint (\code{QSTESTCD == 'CIBIC'})
\item NPI-X Item secondary endpoint (\code{QSTESTCD == 'NPTOT'})
}}
\item{demographics supplemental dataset ('suppdm')}
\item{subject visits ('sv')}
\item{vital signs ('vs')}
}
}
}
This dataset was created following the SDTM Version 2
standard.
}
\seealso{
\link{loadDataADaMSDTM}
}
\author{
Laure Cougnaud
}
\keyword{data}
