% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label.R
\name{formatLabel}
\alias{formatLabel}
\title{concatenate and format text strings to a label, e.g to use for chunk and table/figures}
\usage{
formatLabel(...)
}
\arguments{
\item{...}{string(s) to be concatenated to form label
or data.frame with only one row.
If an unique data.frame is specified, the
different columns are collapsed to form one label.}
}
\value{
string with chunk label
}
\description{
concatenate and format text strings to a label, e.g to use for chunk and table/figures
}
\author{
Laure Cougnaud
}
