% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label.R
\name{formatLabelChunk}
\alias{formatLabelChunk}
\title{Concatenate and format text strings to a chunk label}
\usage{
formatLabelChunk(...)
}
\arguments{
\item{...}{string to be concatenated to form chunk label}
}
\value{
string with chunk label
}
\description{
Concatenate and format text strings to a chunk label
}
\author{
Laure Cougnaud
}
