% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-utility.R
\name{formatLongLabel}
\alias{formatLongLabel}
\title{Format a variable with long labels}
\usage{
formatLongLabel(x, width = 20)
}
\arguments{
\item{x}{character vector with labels to format}

\item{width}{target maximum size. Note: a word
longer that this width won't be split (see \code{\link{strwrap}}).}
}
\value{
vector with formatted labels
}
\description{
This function formats a variable with long labels by wrapping
its elements into multiple lines.
}
\author{
Laure Cougnaud
longLabel <- "This is a very long description of the variable in the dataset"
cat(longLabel)
cat(formatLongLabel(longLabel))
}
