% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label.R
\name{formatTableLabel}
\alias{formatTableLabel}
\title{Concatenate and format text strings to a label of a table}
\usage{
formatTableLabel(...)
}
\arguments{
\item{...}{string to be concatenated to form label}
}
\value{
string with chunk label
}
\description{
This function concatenates and formats 
text strings to a label of a table for \code{bookdown} package
}
\author{
Laure Cougnaud
}
