% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-palettes.R
\name{getColorPalette}
\alias{getColorPalette}
\title{Get a color palette of specified length,
either from a vector of names for the palette, or
from a specified length.}
\usage{
getColorPalette(
  n = NULL,
  x = NULL,
  includeNA = FALSE,
  defaultSettings = clinColors
)
}
\arguments{
\item{n}{Integer of length 1, number of elements in palette.}

\item{x}{Vector with elements used for palette.
If factor, the levels are used, otherwise the unique elements of the vector.
Missing values are automatically removed, excepted if
\code{includeNA} is set to TRUE.}

\item{includeNA}{Logical (FALSE by default), 
should NA elements be retained in the palette in case
\code{x} is specified.}

\item{xPlacebo}{(optional) String with element of \code{x}
containing placebo, for which 'grey' should be used.}

\item{xComparator}{(optional) String with element of \code{x}
containing comparator, for which 'red' should be used.}

\item{type}{String with type of palette used,
either:
\itemize{
\item{'viridis' }{color-blinded friendly palette}
}}

\item{replaceGrey}{Logical, if TRUE (FALSE by default) use
darker grey (rgb(102, 102, 102)).
Kept for backward compatibility of the package.}
}
\value{
Vector of Galapagos colors,
named with the elements in \code{x} if \code{x} is specified.
}
\description{
Get a color palette of specified length,
either from a vector of names for the palette, or
from a specified length.
}
\examples{
# extract longest palette available
getColorPalette(n = 11)
# extract palette for a vector
getColorPalette(x = paste('treatment', 1:4))
# color-blinded palette
getColorPalette(n = 10, type = "viridis")
# possibility to include missing values:
getColorPalette(x = c(NA_character_, "group1"), includeNA = TRUE)
# use more-readable grey color
getColorPalette(n = 10, replaceGrey = TRUE)
}
\author{
Laure Cougnaud
}
