% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{getJavaScriptColumnsIdx}
\alias{getJavaScriptColumnsIdx}
\title{Extract position indices of columns in a data frame and subtract 1}
\usage{
getJavaScriptColumnsIdx(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=getColumnsIdx]{getColumnsIdx}}
  \describe{
    \item{\code{df}}{data frame}
    \item{\code{vars}}{character vector of columns of the data frame}
  }}
}
\description{
This function is used for correspondance between R and javascript column indices
}
\examples{
getJavaScriptColumnsIdx(df = iris, vars = c("Sepal.Width", "Sepal.Length"))
}
\author{
Michela Pasetto
}
