% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitPrintListPlots.R
\name{knitPrintListObjects}
\alias{knitPrintListObjects}
\title{Include a list of objects with possibly different chunk options
in a knitr document (e.g. \code{flextable} object).}
\usage{
knitPrintListObjects(
  xList,
  generalLabel = "objectsList",
  labels = paste0(generalLabel, seq_along(xList)),
  titles = NULL,
  titleLevel = 2,
  printObject = FALSE,
  ...
)
}
\arguments{
\item{xList}{List of objects to print.}

\item{generalLabel}{String with general label for the chunks, 
used to build the \code{labels}.
The labels are constructed as '\code{generalLabel}[i]',
with i the list index.
Only use if \code{labels} is not specified.}

\item{labels}{Character vector with labels,
one for each chunk.}

\item{titles}{Character vector with section titles,
one for each chunk.}

\item{titleLevel}{Integer with level for section header,
1 for top-level section header.}

\item{printObject}{Logical, if TRUE (FALSE by default),
each object within \code{xList} is explicitely printed 
with the \code{\link{print}} function.}

\item{...}{any chunk parameters, will be replicated if necessary,
at the exception of 'results', set to 'asis' and 'echo' set to FALSE
internally.
See \code{\link{knitr}[opts_chunk]} for further details on available options.}
}
\value{
no returned value, a text is printed with chunk content
}
\description{
Note: the chunk in which this function is called should be 
set with the following option: \strong{\code{results = 'asis'}}.
}
\examples{
\dontrun{
# Note: the following code should be included within a chunk of a knitr (e.g. RMarkdown) document
# to include/export a list of figures in the Rmarkdown output

# list of flextable objects
library(flextable)
listTables <- list(flextable(iris), flextable(cars))
knitPrintListObjects(xList = listTables)
}
}
\author{
Laure Cougnaud
}
