% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareTables.R
\name{mergeInputDiff}
\alias{mergeInputDiff}
\title{Custom merge of difference data with input data}
\usage{
mergeInputDiff(diffData, inputData, typeData, colsBy)
}
\arguments{
\item{diffData}{A \code{data.table} object as output from \code{\link{compareDiff}}.}

\item{inputData}{A \code{data.table} object. For instance, the \code{newData} or
the \code{oldData} argument from \code{\link{compareTables}}.}

\item{typeData}{String with type of data, as "new" for \code{newData} or
"old" for \code{oldData}.}

\item{colsBy}{Character vector of columns for doing the merge by.}
}
\value{
A \code{data.table} object.
The \code{inputData} is joined with the columns \code{Comparison type}
and \code{Version} from the \code{diffData} argument.
}
\description{
Custom merge (left join) of difference data with some input data
}
