% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{removeNaCols}
\alias{removeNaCols}
\title{Remove columns with all missing values}
\usage{
removeNaCols(data)
}
\arguments{
\item{data}{data.frame or matrix}
}
\value{
data.frame without columns with all missing values
}
\description{
Remove columns with all missing values
}
\examples{
data <- data.frame(c(NA, NA), B = c("group1", "group2"), C = c(NA, 2))
removeNaCols(data)
}
\author{
Laure Cougnaud
}
