% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{reorderColumns}
\alias{reorderColumns}
\title{Function for reordering columns}
\usage{
reorderColumns(data, vars)
}
\arguments{
\item{data}{A \code{data.frame}}

\item{vars}{Named vector indicating the position 
in the data frame of the specified variable}
}
\value{
The same \code{data.frame} specified in \code{data}, 
with ordered columns.
}
\description{
Function for reordering columns
}
\examples{
someData <- data.frame(
	"Col1" = c(1, 2),
	"Col2" = c(2, 3),
	"Col3" = c(3, 4)
	)
reorderColumns(
	data = someData,
	vars = c("Col3" = 1)
	)
}
