% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{roundCustom}
\alias{roundCustom}
\title{Custom round function, with 'rounding up' strategy 
for rounding off a 5.}
\usage{
roundCustom(x, digits = 0, format = "number")
}
\arguments{
\item{x}{Numeric vector to round.}

\item{digits}{Integer with number of digits to consider, 0 by default.}

\item{format}{String with format for the number. Only 'number' is allowed.
The option 'format = text' has been deprecated.}
}
\value{
Rounded numeric vector.
}
\description{
This function rounds a number up to the nearest number,
for a specified number of digits.
The default R \code{\link{round}} function rounds to the
'even digit' in case of rounding off a 5 
(see 'Details' section in \code{? round}).
This function rounds up to the nearest number in this case,
to mimic a similar rounding strategy used in SAS.
}
\examples{
# numbers are rounded to the closest even number in case of .5 
# with the round 'base' function
round(0.45, 1)
# 'roundCustom' always round to the next highest number in case of .5
roundCustom(0.45, 1)
# rounding is the same for uneven number:
round(0.55, 1)
roundCustom(0.55)
# other examples
round(1.456e-2, digits = 3)
round(1.456e-2, digits = 2)
round(1.456e-2, digits = 1)
}
\author{
stackoverflow question 6461209
}
