% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-palette.R
\docType{data}
\name{shapePaletteNRIND}
\alias{shapePaletteNRIND}
\title{Shape palette for a standard CDISC Normal/Reference
Range Indicator.}
\format{
A named character vector with shape symbol for 
typical Normal Reference Range Indicator variable:
\itemize{
\item{"LOW": }{filled down-pointing arrow (\code{25})}
\item{"NORMAL": }{filled circle (21)}
\item{"HIGH": }{filled up-pointing arrow (\code{24})}
\item{"ABNORMAL": }{diamond (\code{18})}
\item{"UNKNOWN" or 'NA': }{cross (\code{3})}
\item{"NA": }{cross (\code{3})}
}
}
\usage{
shapePaletteNRIND
}
\description{
These symbols should be supported in Windows and Linux.
}
\keyword{datasets}
