% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataADaMCDISCP01}
\alias{dataADaMCDISCP01}
\title{Example of ADaM datasets from the CDISC original Pilot 01 study}
\format{
List of data.frame containing the ADaM
dataset for each selected domain.\cr
Labels for the different variables across datasets
is available via the \code{labelVars} attribute.
}
\source{
Original (and entire) datasets are available in: \cr
\url{https://github.com/phuse-org/phuse-scripts/tree/master/data/adam/cdiscpilot01}
See in particular the \emph{define.xml} file
for further description of the
datasets and variables name.
}
\description{
This contains a subset of the CDISC original Pilot 01 study dataset for:
\itemize{
\item{a selected subset of subjects}
\item{a selected subset of domains: subject-level ('ADSL'),
 adverse event ('adae') and laboratory chemistry ('adlbc') data}
}
}
\seealso{
\link{loadDataADaMSDTM}
}
\author{
Laure Cougnaud
}
\keyword{data}
