% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label.R
\name{formatTableLabel}
\alias{formatTableLabel}
\title{concatenate and format text strings to table label for \code{bookdown} package}
\usage{
formatTableLabel(...)
}
\arguments{
\item{...}{string to be concatenated to form label}
}
\value{
string with chunk label
}
\description{
concatenate and format text strings to table label for \code{bookdown} package
}
\author{
Laure Cougnaud
}
