% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadDataADaMSDTM.R
\name{getLabelVars}
\alias{getLabelVars}
\title{Get label of the variables in SAS dataset(s)}
\usage{
getLabelVars(data, labelVars = NULL, dataList)
}
\arguments{
\item{data}{Data.frame with SAS dataset(s) or list of those.}

\item{labelVars}{(optional) Named character vector with additional labels.}

\item{dataList}{This parameter is deprecated, use \code{data} instead.}
}
\value{
named vector with variable labels
}
\description{
Get label of the variables in SAS dataset(s)
}
\examples{
data(ADaMDataPelican)
# by default extracted from the data (if available)
getLabelVars(ADaMDataPelican)
# extracted from specified labelVars, e.g. to specify custom label for specific variable(s)
labelVarsCustom <- getLabelVars(
  data = ADaMDataPelican, 
  labelVars = c(USUBJID = "Subject identifier for my study")
)
labelVarsCustom["USUBJID"]
}
\author{
Laure Cougnaud
}
