% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-palettes.R
\name{getLinetypePalette}
\alias{getLinetypePalette}
\title{Get a linetype palette of specified length,
either from a vector of names for the palette, or
from a specified length.}
\usage{
getLinetypePalette(
  n = NULL,
  x = NULL,
  xPlacebo = NULL,
  xComparator = NULL,
  includeNA = FALSE
)
}
\arguments{
\item{n}{Integer of length 1, number of elements in palette.}

\item{x}{Vector with elements used for palette.
If factor, the levels are used, otherwise the unique elements of the vector.
Missing values are automatically removed, excepted if
\code{includeNA} is set to TRUE.}

\item{xPlacebo}{(optional) String with element of \code{x}
containing placebo, for which 'longdash' should be used.}

\item{xComparator}{(optional) String with element of \code{x}
containing comparator, for which 'dotted' should be used.}

\item{includeNA}{Logical (FALSE by default), 
should NA elements be retained in the palette in case
\code{x} is specified.}
}
\value{
character vector values with linetype
}
\description{
Note that 7 unique symbols are available at maximum
(replicated if necessary).
}
\examples{
# extract longest linetype palette available
getLinetypePalette(n = 6)
# extract palette for a vector
getLinetypePalette(x = paste('treatment', 1:4))
# include missing
getLinetypePalette(x = c(NA_character_, "group1"), includeNA = TRUE)
}
\author{
Laure Cougnaud
}
